refactor(transport): split native engine into mod/tcp/udp
Addresses review item #7 (part 2/3). No behavior change — pure code move.

The native transport engine (~547 lines) is now an `engine/` module mirroring the
"TCP and UDP are one abstraction at two points" framing the module doc already used:

- mod.rs — shared scaffolding: the SessionKey-keyed session table, connect/listen
  dispatch, write/shutdown/close, RelayTask, and the open/send_frame plumbing.
- tcp.rs — the TCP listener + the bidirectional byte-stream relay loop.
- udp.rs — the UDP listener + the per-flow datagram relay loops + ephemeral-socket bind.

TransportSessions stays the public type at the same path; the split uses module-scoped
privacy (pub(super) for the per-instance entry points).
